requirements.txt
dynaconf==3.1.7
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy-Utils==0.37.8